requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
sentry-sdk==2.57.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
